Added Optional library. Moved files
